refactor(crossposting): serve one crossposter client instead of two
The app had two independently written HTTP clients for the same service,
crossposter.divine.video, distinguished only by "crossposting" vs
"crossposter". Both called GET /connections; each had its own connection
model, its own exception type and its own platform-name mapping. That
split is what let a dead field survive unnoticed in one of them (#7802):
the other model never had it, and nobody noticed the two disagreed.

CrosspostingApiClient absorbs the three job methods and becomes the one
client for the service. CrossposterApiClient, CrossposterConnection and
CrossposterApiException are deleted; CrosspostJob and CrosspostJobStatus
stay, since jobs are a real second concept rather than a duplicate one.
CrosspostingApiException already carried message, statusCode, code and
cause, so nothing is lost — the cubit still switches on e.code and its
mapping is untouched.

Behaviour is deliberately unchanged here. The provider still reads the
raw Keycast session for its token, now through a small adapter; moving
it onto the account-bound reader is the next commit, so it can be
reverted on its own.

Two things worth a reviewer's eye:

Platform typing changes. CrossposterConnection.platform was a String and
kept every value; CrosspostingConnection.platform is an enum and
getConnections skips platforms it does not know. Both cover the same four
today, so nothing changes now, but a fifth server platform would render
in the crosspost sheet before and be skipped after. Taken deliberately:
the settings screen already behaves this way, and two surfaces of one
feature disagreeing is worse than a consistent skip.

Malformed 2xx bodies now keep their diagnostics. The deleted client set
code 'malformed_response' and the FormatException cause; the surviving
_decode dropped both. Carried over, which also improves the settings
endpoints that were silently losing the cause.

Test coverage moves rather than shrinks: the job endpoint cases migrate
into the surviving suite. Its getConnections and close cases were already
duplicated there, more thoroughly.
