+++ mobile/lib/models/crosspost_models.dart
@@ -1,5 +1,5 @@
-// ABOUTME: Wire models for the crossposter service (crossposter.divine.video)
-// ABOUTME: Connections, jobs, job statuses, and the typed API exception
+// ABOUTME: Wire models for crosspost jobs from the crossposter service
+// ABOUTME: Job lifecycle statuses and the per-video job record
 
 /// Lifecycle states reported by the crossposter for a single job.
 enum CrosspostJobStatus {
@@ -34,32 +34,6 @@ enum CrosspostJobStatus {
       this == CrosspostJobStatus.unknown;
 }
 
-/// A platform account link from `GET /connections`.
-class CrossposterConnection {
-  const CrossposterConnection({
-    required this.id,
-    required this.platform,
-    required this.status,
-    this.externalAccountName,
-  });
-
-  factory CrossposterConnection.fromJson(Map<String, dynamic> json) {
-    return CrossposterConnection(
-      id: json['id'] as String? ?? '',
-      platform: json['platform'] as String? ?? '',
-      status: json['status'] as String? ?? '',
-      externalAccountName: json['externalAccountName'] as String?,
-    );
-  }
-
-  final String id;
-  final String platform;
-  final String status;
-  final String? externalAccountName;
-
-  bool get isConnected => status == 'connected';
-}
-
 /// One crosspost job as returned by the crossposter API.
 class CrosspostJob {
   const CrosspostJob({
@@ -91,28 +65,4 @@ class CrosspostJob {
   final String? externalPostUrl;
   final String? errorCode;
   final String? errorMessage;
-}
-
-/// Failure from the crossposter API.
-///
-/// [code] carries the server's machine-readable error code
-/// (`not_owner`, `not_eligible`, `not_connected`, `unauthorized`, ...)
-/// when the response included one.
-class CrossposterApiException implements Exception {
-  const CrossposterApiException(
-    this.message, {
-    this.statusCode,
-    this.code,
-    this.cause,
-  });
-
-  final String message;
-  final int? statusCode;
-  final String? code;
-  final Object? cause;
-
-  @override
-  String toString() =>
-      'CrossposterApiException: $message '
-      '(${statusCode ?? 'no status'}${code == null ? '' : ', $code'})';
 }
